requirements.txt
absl-py==0.11.0
appdirs==1.4.4
astor==0.8.1
audioread==2.1.9
azure-cognitiveservices-speech==1.14.0
beautifulsoup4==4.9.3
boto3==1.16.29
botocore==1.19.29
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
ffmpeg==1.4
Flask==1.1.2
gast==0.4.0
gazpacho==1.1
googleads==3.8.0
grpcio==1.33.2
httplib2==0.18.1
ibm-cloud-sdk-core==1.7.3
ibm-watson==4.7.1
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
inflect==5.0.2
iso8601==0.1.13
itsdangerous==1.1.0
jamo==0.4.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
kiwisolver==1.3.1
librosa==0.7.2
llvmlite==0.31.0
m3u8==0.7.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
nltk==3.5
numba==0.48.0
numpy==1.19.3
oauth2client==4.1.3
packaging==20.4
pandas==1.1.5
Pillow==8.0.1
pooch==1.2.0
praat-parselmouth==0.3.3
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycparser==2.20
pydub==0.24.1
PyJWT==1.7.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2015.7
PyWavelets==1.1.1
pyworld==0.2.4
PyYAML==5.3.1
quote==1.0
regex==2020.11.13
requests==2.25.0
resampy==0.2.2
rsa==4.6
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
spafe==0.1.2
stopit==1.1.1
suds-jurko==0.6
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
tqdm==4.53.0
Unidecode==1.1.1
urllib3==1.26.2
websocket-client==0.48.0
Werkzeug==1.0.1
wincertstore==0.2
zipp==3.4.0
